refactor: faster and cleaner
