+++ pytse_client/__init__.py
@@ -1,5 +1,5 @@
 from .asks_bids import get_asks_and_bids
-from .order_book import get_orderbook
+from .orderbook.order_book import get_orderbook
 from .download import (
     download,
     download_client_types_records,
